backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyyaml==6.0.1
cryptography==41.0.7
slowapi==0.1.9
alembic==1.13.1
python-dotenv==1.0.0
